% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyldap-package.R
\docType{package}
\name{shinyldap-package}
\alias{shinyldap}
\alias{shinyldap-package}
\title{shinyldap: R Shiny Ldap Login Client}
\description{
A dialog box that verify a username and password on an LDAP server.
}
\author{
\strong{Maintainer}: Daniel S \email{suporte@neoage.com.br} (YOUR-ORCID-ID)

}
\keyword{internal}
