% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLdap_demo.R
\name{shinyLdap_demo}
\alias{shinyLdap_demo}
\title{Run a shiny app showing how the ShinyLdap function works.}
\usage{
shinyLdap_demo()
}
\description{
Run a shiny app showing how the ShinyLdap function works.
}
