% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap_functions.R
\name{ldap_login}
\alias{ldap_login}
\title{A ldap_functions Function}
\usage{
ldap_login(input, output, ui_name, modal = FALSE, ldap.url, ldap.dc,
  ldap.filtro = "sAMAccountName", ldap.dominio, ldap.campos,
  label.user = "Usuário", label.pass = "Senha",
  label.button.go = "Login", label.button.cancel = "Cancel",
  label.title = "Shiny LDAP Login", callback.return = function(result)
  { })
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
ldap_login()
}
\keyword{cats}
