% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap_functions.R
\name{ldap_login}
\alias{ldap_login}
\title{A ldap_login function}
\usage{
ldap_login(input, output, ui_name, modal = FALSE, ldap.url, ldap.dc,
  ldap.filtro = "sAMAccountName", ldap.dominio, ldap.campos,
  label.user = "User", label.pass = "Password",
  label.button.go = "Login", label.button.cancel = "Cancel",
  label.button.modal = "Close", label.title = "Shiny LDAP Login",
  show.button.cancel = TRUE, show.button.modal = FALSE,
  callback.return = function(result) { })
}
\arguments{
\item{input}{Shiny input object passed from the server.}

\item{output}{Shiny output object passed from the server.}

\item{ui_name}{the name of the UI output. That is, put \code{uiOutput(ui_name)} where
you want the Login Dialog in \code{ui.R}.}

\item{modal}{boolean (TRUE / FALSE) indicating the form mode.}

\item{ldap.url}{ldap server url}

\item{ldap.dc}{ldap server DC}

\item{ldap.filtro}{'sAMAccountName', # for AD LDAP Server}

\item{ldap.dominio}{ldap server domain}

\item{ldap.campos}{ldap fields to get}

\item{label.user}{user label}

\item{label.pass}{= password label}

\item{label.button.go}{= login button label}

\item{label.button.cancel}{= cancel button label}

\item{label.button.modal}{= modal button label}

\item{label.title}{= Login dialog title}

\item{show.button.cancel}{= Show the cancel button}

\item{show.button.modal}{= Show the modal button}

\item{callback.return}{a function called when the user click a response button. This function can
return a error message.}
}
\description{
This function allows you to verify a username and password on an LDAP server.
}
\examples{
  shinyldap::ldap_login(input, output,
     ui_name = 'ui_login',
     modal = TRUE,
     ldap.url = secrets.ldap.url,
     ldap.dc = secrets.ldap.dc,
     ldap.filtro = secrets.ldap.filtro,
     ldap.dominio = secrets.ldap.dominio,
     ldap.campos = secrets.ldap.campos,
     label.user = 'User',
     label.pass = 'Password',
     label.button.go = 'Login',
     label.button.cancel = 'Cancel',
     label.button.modal = 'Close',
     label.title = 'Shiny LDAP Login',
     show.button.cancel = TRUE,
     show.button.modal = FALSE,
     callback.return = ldap.callback.return)
}
\keyword{ldap}
