% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyLdap-package.R
\docType{package}
\name{ShinyLdap-package}
\alias{ShinyLdap}
\alias{ShinyLdap-package}
\title{ShinyLdap: R Shiny Ldap Login Client (One Line, Title Case)}
\description{
Shiny Ldap Login Client (one paragraph).
}
\author{
\strong{Maintainer}: Daniel S \email{suporte@neoage.com.br} (YOUR-ORCID-ID)

}
\keyword{internal}
